LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/view-timeline-keyframe-boundary-interpolation-expected.txt


FAIL ViewTimeline with timeline offset keyframes outside [0,1] assert_equals: Should have one animation attatched to the view-timeline expected 1 but got 0
